- fix bug where SSE wasn't actually really used at all - improve precision of the SSE version of WriteW - use SSE to calculate the texture wrapping

+++ src/rendering/polyrenderer/drawers/poly_triangle.h
@@ -189,8 +189,8 @@ public:
 	struct Scanline
 	{
 		float W[MAXWIDTH];
-		float U[MAXWIDTH];
-		float V[MAXWIDTH];
+		uint16_t U[MAXWIDTH];
+		uint16_t V[MAXWIDTH];
 		float WorldX[MAXWIDTH];
 		float WorldY[MAXWIDTH];
 		float WorldZ[MAXWIDTH];
